requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
comm==0.2.2
cryptography==42.0.8
debugpy==1.8.2
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
Faker==26.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
gotrue==2.5.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.0.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
postgrest==0.16.8
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
realtime==1.0.6
rich==13.7.1
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
tornado==6.4.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wheel==0.43.0
yarl==1.9.4
